+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/I18n/I18nBase.cs
@@ -20,6 +20,7 @@ namespace KRT.VRCQuestTools.I18n
         internal abstract string YesLabel { get; }
         internal abstract string NoLabel { get; }
         internal abstract string AbortLabel { get; }
+        internal abstract string RemoveLabel { get; }
 
         internal abstract string Maximum { get; }